Revert "Revert "Browserified repository""
