+++ gulpfile.js
@@ -8,11 +8,14 @@ var pug = require("gulp-pug");
 var sass = require("gulp-sass");
 var autoprefixer = require("gulp-autoprefixer");
 var sourcemaps = require("gulp-sourcemaps");
-// var uglify = require('gulp-uglify');
 var notify = require("gulp-notify");
 var browserSync = require("browser-sync").create();
-// var plumber = require('gulp-plumber');
-// var gutil = require("gulp-util");
+var browserify = require('gulp-browserify');
+
+
+let uglify = require('gulp-uglify-es').default;
+let rename = require("gulp-rename");
+
 
 // config
 var paths = {
@@ -20,13 +23,15 @@ var paths = {
     pug: "src/**/*.pug",
     babel: "src/**/*.js",
     sass: "src/**/*.scss",
-    static: "static/**/*"
+    static: "static/**/*",
+    browserify: "src/popup.js"
   },
-  dest: { // 
+  dest: { 
     html: "build",
     js: "build",
     css: "build",
-    static: "build"
+    static: "build",
+    browserify: "build"
   }
 };
 var browsers = "> 1%, last 2 versions, IE >= 9, Firefox ESR"
@@ -51,7 +56,7 @@ gulp.task("pug", function() {
 });
 
 gulp.task("babel", function() {
-  return gulp.src(paths.src.babel)
+  return gulp.src([paths.src.babel, "!src/**/popup.js"])
     // .pipe(sourcemaps.init())
     .pipe(babel({
       presets: [
@@ -101,26 +106,29 @@ gulp.task("static", function() {
     }));
 });
 
-// function makeWatcher(fileType) {
-//   console.log(fileType, paths.src[fileType]);
-//   gulp.watch(paths.src[fileType], [fileType])
-//     // Special handler for deleting files
-//     .on("change", (event) => {
-//       if (event.type === 'deleted') {
-//         var srcPath = path.relative(path.resolve(paths.src[fileType]), event.path);
-//         var destPath = path.resolve(paths.dest[fileType], srcPath);
-//         del.sync(destPath);
-//         browserSync.reload()
-//       }
-//     });
-// }
+gulp.task("scripts", function() {
+    // Single entry point to browserify 
+    return gulp.src('src/popup.js')
+        .pipe(browserify({
+          insertGlobals : true,
+          debug : false
+        }))
+        .pipe(browserSync.stream())
+        .pipe(uglify())
+        .pipe(gulp.dest(paths.dest.browserify))
+        .pipe(notify({
+          title: "Success",
+          message: "Browserfied: <%= file.relative %>"
+        }));
+});
+
 
 gulp.task("clean", function(done) {
   return del("build/**/*", done);
 });
 
-gulp.task("watch", function() {
-  // ["pug", "babel", "sass", "static"].forEach(makeWatcher);
+gulp.task("watch", function() {  
+  gulp.watch(paths.src.browserify, gulp.task("scripts"));
   gulp.watch(paths.src.pug, gulp.task("pug"));
   gulp.watch(paths.src.babel, gulp.task("babel"));
   gulp.watch(paths.src.sass, gulp.task("sass"));
@@ -137,7 +145,7 @@ gulp.task("sync", function() {
 
 gulp.task("build",
   gulp.series("clean",
-    gulp.parallel("pug", "babel", "sass", "static")));
+    gulp.parallel("pug", "babel", "sass", "static", "scripts")));
 
 gulp.task("default", 
   gulp.series("build", "watch"));
