+++ package.json
@@ -23,10 +23,14 @@
     "gulp": "github:gulpjs/gulp#4.0",
     "gulp-autoprefixer": "^3.1.1",
     "gulp-babel": "^6.1.2",
+    "gulp-browserify": "^0.5.1",
     "gulp-notify": "^2.2.0",
     "gulp-pug": "^3.2.0",
+    "gulp-rename": "^1.2.2",
     "gulp-sass": "^3.1.0",
-    "gulp-sourcemaps": "^2.4.0"
+    "gulp-sourcemaps": "^2.4.0",
+    "gulp-uglify-es": "^0.1.3",
+    "uglify-es": "^3.0.28"
   },
   "dependencies": {
     "body-parser": "^1.17.2",
