+++ static/manifest.json
@@ -2,7 +2,7 @@
   "manifest_version": 2,
   "name": "Timelime",
   "description": "",
-  "version": "1.0",
+  "version": "0.1.0",
   "background": {
     "scripts": ["background.js"],
     "persistent": false
@@ -21,5 +21,10 @@
     "tabs",
     "<all_urls>",
     "http://127.0.0.1/*"
-  ]
+  ],
+  "icons": { 
+    "16": "icon16.png",           
+    "48": "icon48.png",
+    "128": "icon128.png" 
+  }
 }
